pyproject.toml
[tool.poetry]
name = "binance"
version = "0.1.0"
description = ""
authors = ["Vas <vas@va.va>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-binance = "^1.0.17"
pandas = "^1.5.3"
statsmodels = "^0.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
